==> 1.1 Articulos_FecUltCom.ipynb <==
# Importar librerías
import pandas as pd 
import os

# Establecer directorio de trabajo
os.chdir("/Users/adrianamedrano/Documents/Ariana/Residencias/PrototipoResidencias/DataSets")

# Leer archivos CSV en dataframes
df_articulos = pd.read_csv("Articulos.txt", encoding="latin1")  
df_profd = pd.read_csv("Profd.txt", encoding="latin1")
df_prof = pd.read_csv("Prof.txt", encoding="latin1")

# Unir dataframes por clave común 
df_merge = df_articulos.merge(df_profd, on="Claart", how="left").merge(df_prof, on="Numfac", how="left")

# Obtener fecha máxima por articulo
df_fecha_max = df_merge.groupby("Claart")["Fecfac"].max().reset_index() 

# Establecer indices
df_articulos = df_articulos.set_index("Claart")
df_fecha_max = df_fecha_max.set_index("Claart")

# Llenar valores nulos en df_articulos con fecha máxima  
df_articulos["FecUltCom"] = df_articulos["FecUltCom"].fillna(df_fecha_max["Fecfac"])

# Reestablecer indice a columna
df_articulos = df_articulos.reset_index() 

# Llenar valores nulos restantes
df_articulos["FecUltCom"].fillna("27/07/16", inplace=True) 

# Guardar dataframe
df_articulos.to_csv("/Users/adrianamedrano/Documents/Ariana/Residencias/PrototipoResidencias/DataSets_Codigo/Articulos.txt", index=False, encoding="latin1")
