==> 1.4 Proveedores.ipynb <==
# Importar librerías
import pandas as pd
import os

# Establecer directorio de trabajo
os.chdir("/Users/adrianamedrano/Documents/Ariana/Residencias/PrototipoResidencias/DataSets")

# Leer archivos CSV 
df_proveedores = pd.read_csv("Proveedores.txt", encoding="latin1")
df_prof = pd.read_csv("Prof.txt", encoding="latin1")

# Convertir columnas de fecha a datetime
df_prof["Fecfac"] = pd.to_datetime(df_prof["Fecfac"], format="%d/%m/%y")
df_proveedores["Fecultcom"] = pd.to_datetime(df_proveedores["Fecultcom"], format="%d/%m/%y")

# Obtener fecha máxima de factura por proveedor
fecha_mas_reciente = df_prof.groupby("Clapro")["Fecfac"].max().reset_index()

# Unir dataframes y llenar fecha última compra con fecha máxima
df_proveedores = df_proveedores.merge(fecha_mas_reciente, on="Clapro", how="left")
df_proveedores["Fecultcom"].fillna(df_proveedores["Fecfac"], inplace=True) 

# Eliminar columna utilizada para llenado
df_proveedores.drop("Fecfac", axis=1, inplace=True)

# Formatear fechas a string
df_proveedores["Fecultcom"] = df_proveedores["Fecultcom"].dt.strftime("%d/%m/%y")
df_prof["Fecfac"] = df_prof["Fecfac"].dt.strftime("%d/%m/%y")

# Eliminar filas con fecha nula
df_proveedores = df_proveedores.dropna(subset=["Fecultcom"])

# Guardar dataframe
df_proveedores.to_csv('/Users/adrianamedrano/Documents/Ariana/Residencias/PrototipoResidencias/DataSets_Codigo/Proveedores.txt', index=False, encoding="latin1")
