==> 2.1 C_Articulos.ipynb <==
# Importar librerías
import pandas as pd
import os

# Establecer directorio de trabajo 
os.chdir("/Users/adrianamedrano/Documents/Ariana/Residencias/PrototipoResidencias/DataSets")

# Leer archivo CSV de artículos
df_articulos = pd.read_csv("Articulos.txt", encoding="latin1")

# Convertir columnas de fecha a datetime
df_articulos['FecUltCom'] = pd.to_datetime(df_articulos['FecUltCom'], format="%d/%m/%y")

# Separar fecha última compra en día, mes y año
df_articulos.insert(df_articulos.columns.get_loc('FecUltCom') + 1, 'FUCDia', df_articulos['FecUltCom'].dt.day)  
df_articulos.insert(df_articulos.columns.get_loc('FUCDia') + 1, 'FUCMes', df_articulos['FecUltCom'].dt.month)
df_articulos.insert(df_articulos.columns.get_loc('FUCMes') + 1, 'FUCAño', df_articulos['FecUltCom'].dt.year)

# Convertir fecha última compra a string 
df_articulos["FecUltCom"] = df_articulos["FecUltCom"].dt.strftime("%d/%m/%y")

# Repetir proceso para fecha última venta 
df_articulos['FecUltven'] = pd.to_datetime(df_articulos['FecUltven'], format="%d/%m/%y")

df_articulos.insert(df_articulos.columns.get_loc('FecUltven') + 1, 'FUvDia', df_articulos['FecUltven'].dt.day)
df_articulos.insert(df_articulos.columns.get_loc('FUvDia') + 1, 'FUvMes', df_articulos['FecUltven'].dt.month) 
df_articulos.insert(df_articulos.columns.get_loc('FUvMes') + 1, 'FUvAño', df_articulos['FecUltven'].dt.year)

df_articulos["FecUltven"] = df_articulos["FecUltven"].dt.strftime("%d/%m/%y")

# Guardar archivo CSV
df_articulos.to_csv('/Users/adrianamedrano/Documents/Ariana/Residencias/PrototipoResidencias/DataSets_Codigo/Articulos.txt', index=False, encoding="latin1")
