==> 3.3 Mer_clied.ipynb <==
# Importar librerías
import pandas as pd
import os

# Establecer directorio de trabajo
os.chdir("/Users/adrianamedrano/Documents/Ariana/Residencias/PrototipoResidencias/DataSets")

# Leer archivos CSV
df_clie = pd.read_csv('clie.txt', encoding='latin1')
df_clied = pd.read_csv('clied.txt', encoding='latin1')

# Mergear dataframes por campo Nummov 
df_clied = pd.merge(df_clied, df_clie[['Nummov','Fecncre','Estado']],
                    on='Nummov', 
                    how='left',
                    suffixes=('','_clie'))

# Filtrar por Estado = 'yes'
df_clied = df_clied[df_clied['Estado_clie'] == 'yes'][['Claart', 'Fecncre', 'Cant']]

# Guardar dataframe 
df_clied.to_csv('/Users/adrianamedrano/Documents/Ariana/Residencias/PrototipoResidencias/DataSets_Codigo/Mer_clied.txt', index=False, encoding="latin1")
